Fix -Wconversion errors

+++ include/TsParser.h
@@ -88,7 +88,7 @@ public:
      * @param tsPacketInfo Input packet inforamtion
      * @param table_id Collected table id
      */
-    void collectTable(const uint8_t* tsPacket, const TsPacketInfo& tsPacketInfo, uint8_t& table_id);
+    void collectTable(const uint8_t* tsPacket, const TsPacketInfo& tsPacketInfo, int& table_id);
     /*!
      * Parses PSI table
      * @param packet
@@ -139,7 +139,7 @@ public:
 
 private:
     // TODO maybe 1 parser per pid?
-    std::map<uint16_t, ByteVector> mSectionBuffer;
+    std::map<int, ByteVector> mSectionBuffer;
     std::map<uint16_t, int> mSectionLength;
     std::map<uint16_t, int> mTableId;
     std::map<uint16_t, int> mReadSectionLength;
